[FIX] Check setup.py for local package installation

+++ scripts/setup_venv.sh
@@ -18,7 +18,9 @@ else
     install_dependencies requirements.txt
 
     echo "Installing local package"
-    python -m pip install --upgrade -e .
+    if [ -e setup.py ]; then
+        python -m pip install --upgrade -e .
+    fi
 
     echo "Installing modules"
     for package in $(ls -d */); do pushd $package; if [ -e setup.py ]; then python -m pip install --upgrade -e .; fi; popd; done
